fix: read /proc for max pid on linux
`/proc/sys/kernel/pid_max` will provide the maximum pid on linux.
This patch reads that value to use in `validate_pid`, and then
turns it around to verify that the pid read from the pidfile is
valid. If that file is not present, or cannot be read, the
previous "guess" will be used.

Thanks to @linsomniac for the tip.